view artist info

+++ app/src/main/java/com/revosleap/bxplayer/models/Song.kt
@@ -9,6 +9,6 @@ class Song : Serializable {
     var albumName: String? = ""
     var artistId: Int = 0
     var songYear: Int = 0
-    var duration: Int = 0
+    var duration: Long = 0L
     var trackNumber:Int=0
 }
